quick n dirty inky serverver w/quant/resize
